clean up and improve receive screen
