+++ app/components/receive-screen/index.ts
@@ -1,5 +1,17 @@
 import WalletBottomSheet from "./WalletBottomSheet"
 import ReceiveTypeBottomSheet from "./ReceiveTypeBottomSheet"
 import AmountNote from "./AmountNote"
+import OnChainCharge from "./OnChainCharge"
+import InvoiceInfo from "./InvoiceInfo"
+import WalletReceiveTypeTabs from "./WalletReceiveTypeTabs"
+import Header from "./Header"
 
-export { WalletBottomSheet, ReceiveTypeBottomSheet, AmountNote }
+export {
+  WalletBottomSheet,
+  ReceiveTypeBottomSheet,
+  AmountNote,
+  OnChainCharge,
+  InvoiceInfo,
+  WalletReceiveTypeTabs,
+  Header,
+}
