+++ app/screens/receive-bitcoin-screen/receive-screen.tsx
@@ -1,144 +1,61 @@
 import React, { useEffect, useState } from "react"
-import { Share, TouchableOpacity, View } from "react-native"
-import { useIsFocused, useNavigation } from "@react-navigation/native"
+import { useNavigation } from "@react-navigation/native"
 import Clipboard from "@react-native-clipboard/clipboard"
-import Icon from "react-native-vector-icons/Ionicons"
 import { useI18nContext } from "@app/i18n/i18n-react"
-import { makeStyles, Text, useTheme } from "@rneui/themed"
+import { makeStyles } from "@rneui/themed"
 import { useReceiveBitcoin } from "./use-receive-bitcoin"
 import { useAppSelector } from "@app/store/redux"
-import nfcManager from "react-native-nfc-manager"
 import { useAppConfig } from "@app/hooks"
-import moment from "moment"
-import messaging from "@react-native-firebase/messaging"
-import { usePersistentStateContext } from "@app/store/persistent-state"
+
+// breez
+import {
+  addEventListener,
+  removeEventListener,
+  SdkEvent,
+  SdkEventVariant,
+} from "@breeztech/react-native-breez-sdk-liquid"
 
 // components
 import { Screen } from "@app/components/screen"
 import { QRView } from "./qr-view"
-import { AmountInput } from "@app/components/amount-input"
-import { NoteInput } from "@app/components/note-input"
 import { SetLightningAddressModal } from "@app/components/set-lightning-address-modal"
 import { withMyLnUpdateSub } from "./my-ln-updates-sub"
-import { CustomIcon } from "@app/components/custom-icon"
 import { ModalNfc } from "@app/components/modal-nfc"
 import {
   AmountNote,
-  ReceiveTypeBottomSheet,
-  WalletBottomSheet,
+  Header,
+  InvoiceInfo,
+  OnChainCharge,
+  WalletReceiveTypeTabs,
 } from "@app/components/receive-screen"
 
 // gql
 import { WalletCurrency, useWalletsQuery } from "@app/graphql/generated"
-import { useIsAuthed } from "@app/graphql/is-authed-context"
-import { useApolloClient } from "@apollo/client"
-
-// breez-sdk
-import { paymentEvents } from "@app/utils/breez-sdk-liquid"
-
-// utils
-import { addDeviceToken, requestNotificationPermission } from "@app/utils/notifications"
-import { testProps } from "../../utils/testProps"
 
 // types
-import { Invoice, InvoiceType, PaymentRequestState } from "./payment/index.types"
+import { Invoice, PaymentRequestState } from "./payment/index.types"
+
+// store
+import { usePersistentStateContext } from "@app/store/persistent-state"
 
 const ReceiveScreen = () => {
-  const { userData } = useAppSelector((state) => state.user)
-  const { colors } = useTheme().theme
+  const navigation = useNavigation()
   const styles = useStyles()
   const { LL } = useI18nContext()
-  const navigation = useNavigation()
-  const client = useApolloClient()
-
   const { appConfig } = useAppConfig()
-  const isAuthed = useIsAuthed()
-  const isFocused = useIsFocused()
-
-  const lnAddressHostname = appConfig.galoyInstance.lnAddressHostname
-
+  const { userData } = useAppSelector((state) => state.user)
+  const { data, loading, error } = useWalletsQuery()
   const request = useReceiveBitcoin()
 
   const { persistentState } = usePersistentStateContext()
   const [displayReceiveNfc, setDisplayReceiveNfc] = useState(false)
+  const [updatedPaymentState, setUpdatedPaymentState] = React.useState<string>()
+  const [lnurlp, setLnurlp] = useState<string>()
+  const [breezListenerId, setBreezListenerId] = useState<string>()
 
-  // query
-  const { data, loading, error } = useWalletsQuery()
+  const lnAddressHostname = appConfig.galoyInstance.lnAddressHostname
   const wallets: any = data?.me?.defaultAccount.wallets
   const usdWallet = wallets?.find((el: any) => el.walletCurrency === WalletCurrency.Usd)
-
-  const nfcText = LL.ReceiveScreen.nfc()
-  useEffect(() => {
-    ;(async () => {
-      if (
-        request?.type === "Lightning" &&
-        request?.state === "Created" &&
-        (await nfcManager.isSupported())
-      ) {
-        console.log("DEBUG:", request.receiveViaNFC)
-        navigation.setOptions({
-          headerRight: () => (
-            <TouchableOpacity
-              style={styles.nfcIcon}
-              onPress={() => setDisplayReceiveNfc(true)}
-            >
-              <Text type="p2">{nfcText}</Text>
-              <CustomIcon name="nfc" color={colors.black} />
-            </TouchableOpacity>
-          ),
-        })
-      } else {
-        navigation.setOptions({ headerRight: () => <></> })
-      }
-    })()
-    // Disable exhaustive-deps because styles.nfcIcon was causing an infinite loop
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [nfcText, colors.black, navigation, request?.state, request?.type])
-
-  useEffect(() => {
-    let timeout: NodeJS.Timeout
-    if (isAuthed && isFocused) {
-      const WAIT_TIME_TO_PROMPT_USER = 5000
-      timeout = setTimeout(
-        requestNotificationPermission, // no op if already requested
-        WAIT_TIME_TO_PROMPT_USER,
-      )
-    }
-    return () => timeout && clearTimeout(timeout)
-  }, [isAuthed, isFocused])
-
-  useEffect(() => {
-    if (persistentState.isAdvanceMode) {
-      switch (request?.type) {
-        case Invoice.OnChain:
-          navigation.setOptions({ title: LL.ReceiveScreen.receiveViaOnchain() })
-          break
-        case Invoice.Lightning:
-          navigation.setOptions({ title: LL.ReceiveScreen.receiveViaInvoice() })
-          break
-        case Invoice.PayCode:
-          navigation.setOptions({ title: LL.ReceiveScreen.receiveViaPaycode() })
-      }
-    } else {
-      navigation.setOptions({ title: LL.ReceiveScreen.receive() })
-    }
-  }, [request?.type])
-
-  const [updatedPaymentState, setUpdatedPaymentState] = React.useState<
-    string | undefined
-  >(undefined)
-
-  useEffect(() => {
-    if (request) {
-      // Subscribe to the "paymentSuccess" event.
-      paymentEvents.once("invoicePaid", handleInvoicePaid)
-
-      // Clean up: unsubscribe to prevent memory leaks.
-      return () => {
-        paymentEvents.off("invoicePaid", handleInvoicePaid)
-      }
-    }
-  }, [request])
 
   useEffect(() => {
     if (
@@ -150,179 +67,109 @@ const ReceiveScreen = () => {
     }
   }, [request?.state, updatedPaymentState, navigation])
 
-  // notification permission
   useEffect(() => {
-    let timeout: NodeJS.Timeout
-    if (isAuthed && isFocused && client) {
-      const WAIT_TIME_TO_PROMPT_USER = 5000
-      timeout = setTimeout(
-        async () => {
-          const result = await requestNotificationPermission()
-          if (
-            result === messaging.AuthorizationStatus.PROVISIONAL ||
-            result === messaging.AuthorizationStatus.AUTHORIZED
-          ) {
-            await addDeviceToken(client)
-          }
-        }, // no op if already requested
-        WAIT_TIME_TO_PROMPT_USER,
-      )
-    }
-    return () => timeout && clearTimeout(timeout)
-  }, [isAuthed, isFocused, client])
+    if (persistentState.isAdvanceMode && !breezListenerId) addBreezEventListener()
+    return removeBreezEventListener
+  }, [persistentState.isAdvanceMode, breezListenerId])
 
-  const handleInvoicePaid = () => {
-    if (request) {
-      request.state = PaymentRequestState.Paid
-      if (request?.state === PaymentRequestState.Paid) {
+  useEffect(() => {
+    if (
+      request &&
+      request.type === "PayCode" &&
+      request.receivingWalletDescriptor.currency === "USD" &&
+      Boolean(usdWallet?.lnurlp) &&
+      Boolean(userData.username)
+    ) {
+      setLnurlp(usdWallet?.lnurlp)
+    } else {
+      setLnurlp(undefined)
+    }
+  }, [request, usdWallet?.lnurlp])
+
+  const addBreezEventListener = async () => {
+    const listenerId = await addEventListener((e: SdkEvent) => {
+      if (
+        e.type === SdkEventVariant.PAYMENT_WAITING_CONFIRMATION ||
+        e.type === SdkEventVariant.PAYMENT_SUCCEEDED
+      ) {
         setUpdatedPaymentState(PaymentRequestState.Paid)
       }
-    }
+    })
+    setBreezListenerId(listenerId)
   }
 
-  if (!request) return <></>
-
-  const OnChainCharge =
-    request.feesInformation?.deposit.minBankFee &&
-    request.feesInformation?.deposit.minBankFeeThreshold &&
-    request.type === Invoice.OnChain ? (
-      <View style={styles.onchainCharges}>
-        <Text type="p4">
-          {LL.ReceiveScreen.fees({
-            minBankFee: request.feesInformation?.deposit.minBankFee,
-            minBankFeeThreshold: request.feesInformation?.deposit.minBankFeeThreshold,
-          })}
-        </Text>
-      </View>
-    ) : undefined
-
-  const isReady = request.state !== PaymentRequestState.Loading
-
-  const lnurlp = usdWallet?.lnurlp || ""
-  const useLnurlp =
-    request.type === "PayCode" &&
-    request.receivingWalletDescriptor.currency === "USD" &&
-    Boolean(lnurlp) &&
-    Boolean(userData.username)
+  const removeBreezEventListener = () => {
+    if (breezListenerId) {
+      removeEventListener(breezListenerId)
+    }
+  }
 
   const handleCopy = () => {
-    if (request.type === "PayCode") {
-      Clipboard.setString(`${request.username}@${lnAddressHostname}`)
-    } else {
-      if (request.copyToClipboard) {
-        request.copyToClipboard()
+    if (request) {
+      if (request.type === "PayCode") {
+        Clipboard.setString(`${request.username}@${lnAddressHostname}`)
+      } else {
+        if (request.copyToClipboard) {
+          request.copyToClipboard()
+        }
       }
     }
   }
 
-  const handleShare = async () => {
-    if (useLnurlp) {
-      const result = await Share.share({ message: lnurlp })
-    } else {
-      if (request.share) {
-        request.share()
-      }
-    }
+  if (request) {
+    return (
+      <>
+        <Header request={request} setDisplayReceiveNfc={setDisplayReceiveNfc} />
+        <Screen
+          preset="scroll"
+          keyboardOffset="navigationHeader"
+          keyboardShouldPersistTaps="handled"
+          style={styles.screenStyle}
+        >
+          <WalletReceiveTypeTabs request={request} />
+          <QRView
+            type={request.info?.data?.invoiceType || Invoice.OnChain}
+            getFullUri={!!lnurlp ? lnurlp : request.info?.data?.getFullUriFn}
+            loading={!!lnurlp ? loading : request.state === PaymentRequestState.Loading}
+            completed={
+              updatedPaymentState === PaymentRequestState.Paid ||
+              request.state === PaymentRequestState.Paid
+            }
+            err={
+              request.state === PaymentRequestState.Error || (!!lnurlp && error)
+                ? LL.ReceiveScreen.error()
+                : ""
+            }
+            style={styles.qrView}
+            expired={request.state === PaymentRequestState.Expired}
+            regenerateInvoiceFn={request.regenerateInvoice}
+            copyToClipboard={handleCopy}
+            isPayCode={request.type === Invoice.PayCode}
+            canUsePayCode={!!lnurlp || request.canUsePaycode}
+            toggleIsSetLightningAddressModalVisible={
+              request.toggleIsSetLightningAddressModalVisible
+            }
+          />
+          <InvoiceInfo request={request} lnurlp={lnurlp} handleCopy={handleCopy} />
+          <AmountNote request={request} />
+          <OnChainCharge request={request} />
+          <SetLightningAddressModal
+            isVisible={request.isSetLightningAddressModalVisible}
+            toggleModal={request.toggleIsSetLightningAddressModalVisible}
+          />
+          <ModalNfc
+            isActive={displayReceiveNfc}
+            setIsActive={setDisplayReceiveNfc}
+            settlementAmount={request.settlementAmount}
+            receiveViaNFC={request.receiveViaNFC}
+            onPaid={() => setUpdatedPaymentState(PaymentRequestState.Paid)}
+          />
+        </Screen>
+      </>
+    )
+  } else {
+    return null
   }
-
-  const onChangeWallet = (id: WalletCurrency) => {
-    if (isReady) {
-      if (id === "BTC" && request.type === "PayCode") {
-        request.setType("Lightning")
-      }
-      request.setReceivingWallet(id)
-    }
-  }
-
-  return (
-    <>
-      <Screen
-        preset="scroll"
-        keyboardOffset="navigationHeader"
-        keyboardShouldPersistTaps="handled"
-        style={styles.screenStyle}
-      >
-        {persistentState.isAdvanceMode && (
-          <View style={{ flexDirection: "row", marginBottom: 10 }}>
-            <WalletBottomSheet
-              currency={request.receivingWalletDescriptor.currency}
-              disabled={request.state === PaymentRequestState.Loading}
-              onChange={onChangeWallet}
-            />
-            <View style={{ width: 10 }} />
-            <ReceiveTypeBottomSheet
-              currency={request.receivingWalletDescriptor.currency}
-              type={request.type}
-              disabled={request.state === PaymentRequestState.Loading}
-              onChange={(id) => isReady && request.setType(id as InvoiceType)}
-            />
-          </View>
-        )}
-        <QRView
-          type={request.info?.data?.invoiceType || Invoice.OnChain}
-          getFullUri={useLnurlp ? lnurlp : request.info?.data?.getFullUriFn}
-          loading={useLnurlp ? loading : request.state === PaymentRequestState.Loading}
-          completed={
-            updatedPaymentState === PaymentRequestState.Paid ||
-            request.state === PaymentRequestState.Paid
-          }
-          err={
-            request.state === PaymentRequestState.Error || (useLnurlp && error)
-              ? LL.ReceiveScreen.error()
-              : ""
-          }
-          style={styles.qrView}
-          expired={request.state === PaymentRequestState.Expired}
-          regenerateInvoiceFn={request.regenerateInvoice}
-          copyToClipboard={handleCopy}
-          isPayCode={request.type === Invoice.PayCode}
-          canUsePayCode={useLnurlp || request.canUsePaycode}
-          toggleIsSetLightningAddressModalVisible={
-            request.toggleIsSetLightningAddressModalVisible
-          }
-        />
-
-        {request.state !== PaymentRequestState.Loading &&
-          request.info?.data?.invoiceType === Invoice.Lightning && (
-            <View style={styles.invoiceDetails}>
-              <Text color={colors.grey2}>
-                {request.state === PaymentRequestState.Expired
-                  ? LL.ReceiveScreen.invoiceHasExpired()
-                  : `Valid for ${moment(request.info.data.expiresAt).fromNow(true)}`}
-              </Text>
-            </View>
-          )}
-        {request.state !== PaymentRequestState.Loading &&
-          request.readablePaymentRequest && (
-            <View style={styles.extraDetails}>
-              <TouchableOpacity onPress={handleCopy}>
-                <Text {...testProps("readable-payment-request")}>
-                  {request.readablePaymentRequest}
-                </Text>
-              </TouchableOpacity>
-              <TouchableOpacity onPress={handleShare} style={styles.shareInvoice}>
-                <Icon color={colors.grey2} name="share-outline" size={20} />
-              </TouchableOpacity>
-            </View>
-          )}
-
-        <AmountNote request={request} />
-        {OnChainCharge}
-
-        <SetLightningAddressModal
-          isVisible={request.isSetLightningAddressModalVisible}
-          toggleModal={request.toggleIsSetLightningAddressModalVisible}
-        />
-        <ModalNfc
-          isActive={displayReceiveNfc}
-          setIsActive={setDisplayReceiveNfc}
-          settlementAmount={request.settlementAmount}
-          receiveViaNFC={request.receiveViaNFC}
-          onPaid={() => setUpdatedPaymentState(PaymentRequestState.Paid)}
-        />
-      </Screen>
-    </>
-  )
 }
 
 const useStyles = makeStyles(({ colors }) => ({
@@ -331,77 +178,8 @@ const useStyles = makeStyles(({ colors }) => ({
     paddingVertical: 12,
     flexGrow: 1,
   },
-  tabRow: {
-    flexDirection: "row",
-    flexWrap: "nowrap",
-    justifyContent: "center",
-    marginTop: 12,
-  },
-  usdActive: {
-    backgroundColor: colors.green,
-    borderRadius: 7,
-    justifyContent: "center",
-    alignItems: "center",
-    width: 150,
-    height: 30,
-    margin: 5,
-  },
-  btcActive: {
-    backgroundColor: colors.primary,
-    borderRadius: 7,
-    justifyContent: "center",
-    alignItems: "center",
-    width: 150,
-    height: 30,
-    margin: 5,
-  },
-  inactiveTab: {
-    backgroundColor: colors.grey3,
-    borderRadius: 7,
-    justifyContent: "center",
-    alignItems: "center",
-    width: 150,
-    height: 30,
-    margin: 5,
-  },
   qrView: {
     marginBottom: 10,
-  },
-  receivingWalletPicker: {
-    marginBottom: 10,
-  },
-  invoiceTypePicker: {
-    marginBottom: 10,
-  },
-
-  extraDetails: {
-    flexDirection: "row",
-    justifyContent: "center",
-    alignItems: "center",
-    marginBottom: 10,
-  },
-  invoiceDetails: {
-    alignItems: "center",
-    marginBottom: 10,
-  },
-  shareInvoice: {
-    marginLeft: 5,
-  },
-  onchainCharges: { marginTop: 10, alignItems: "center" },
-  warning: {
-    fontSize: 12,
-    color: colors.warning,
-    marginBottom: 10,
-  },
-  nfcIcon: {
-    marginTop: -1,
-    marginRight: 14,
-    padding: 8,
-    display: "flex",
-    flexDirection: "row",
-    columnGap: 4,
-    backgroundColor: colors.grey5,
-    borderRadius: 4,
   },
 }))
